pyproject.toml
[project]
name = "eonreader"
version = "0.1.0"
description = ""
authors = [
    {name = "sergiubolos",email = "bogdan.bolos@rebeldot.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pypdf2",
    "pdfplumber",
    "pymupdf",
    "pdf2image",
    "pytesseract",
    "openai",
    "langchain",
    "python-dotenv",
    "streamlit",
    "pandas",
    "matplotlib"
]

[tool.poetry]
packages = [
    { include = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
